requirements.txt
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
folium==0.19.5
google-generativeai==0.3.2
geopy==2.4.1
python-dateutil==2.8.2
httpx[http2]==0.27.0
polyline==2.0.0